+++ app/src/Playground/index.tsx
@@ -1,25 +1,32 @@
 /* eslint-disable no-console */
 
-import { useForm, useFieldArray } from "react-cool-form";
+import { useForm } from "react-cool-form";
 
 export default () => {
-  const { form } = useForm({
-    defaultValues: { foo: [{ name: "test-1" }] },
-    onSubmit: (values, { getState }) => console.log("onSubmit: ", getState()),
+  const { form, focus } = useForm({
+    defaultValues: { foo: [{ a: "test-1", b: "test-1" }] },
+    onSubmit: (values) => console.log("onSubmit: ", values),
   });
-  const [fields, { remove }] = useFieldArray("foo", {
-    // defaultValue: [{ name: "test-1" }],
-  });
+  const [fields, { push }] = useFieldArray("foo");
 
   return (
     <form ref={form}>
-      {fields.map((fieldName) => (
-        <input key={fieldName} name={`${fieldName}.name`} />
+      {fields.map((name) => (
+        <div key={name}>
+          <input name={`${name}.a`} />
+          <input name={`${name}.b`} />
+        </div>
       ))}
-      <button type="button" onClick={() => remove(0)}>
-        Remove
+      <button
+        type="button"
+        onClick={() => {
+          push({ a: "test-2", b: "test-2" });
+          focus(`foo[${fields.length}]`, 300);
+          // focus(`foo[${fields.length}].b`, 300);
+        }}
+      >
+        Push
       </button>
-      <input type="submit" />
     </form>
   );
 };